[REFACT] : made Room component and moved its related code into the component from Rooms.tsx - Room 모듈화
